fix: use double instead of decimal and add db migration

+++ ECommerce/ECommerce.Services.CouponAPI/Models/Coupon.cs
@@ -9,7 +9,7 @@ namespace ECommerce.Services.CouponAPI.Models
         [Required]
         public string CouponCode { get; set; }
         [Required]
-        public decimal DiscountAmount { get; set; }
+        public double DiscountAmount { get; set; }
         public int MinAmount { get; set; }
     }
 }
